Updated example configuration for auth0 usage
Note: These are already the correct values for dev&prod usage. There is no need to hide these values because the security does not come from the obscurity of our auth domain etc., but from the username & password that the user enters. The client ID, domain etc. are visible in the generated login URL anyways.